Remove tmp_dir only if the output dir is not in s3 (#320)
* Remove tmp_dir only if we are local

* Skip removal of tmp_dir if filesystem is s3

* fix black

+++ img2dataset/main.py
@@ -255,7 +255,9 @@ def download(
         max_shard_retry,
     )
     logger_process.join()
-    fs.rm(tmp_dir, recursive=True)
+
+    if not hasattr(fs, "s3"):
+        fs.rm(tmp_dir, recursive=True)
 
 
 def main():
